optimized buffer for data(only 15GB used) and included size printouts

+++ buffer.py
@@ -7,10 +7,10 @@ class ReplayBuffer():
     def __init__(self, max_size, input_shape, n_actions, device='cpu'):
         self.mem_size = max_size
         self.mem_ctr = 0
-        self.state_memory = np.zeros((self.mem_size, *input_shape))
-        self.next_state_memory = np.zeros((self.mem_size, *input_shape))
-        self.action_memory = np.zeros((self.mem_size, n_actions))
-        self.reward_memory = np.zeros(self.mem_size)
+        self.state_memory = np.zeros((self.mem_size, *input_shape), dtype=np.uint8)
+        self.next_state_memory = np.zeros((self.mem_size, *input_shape), dtype=np.uint8)
+        self.action_memory = np.zeros((self.mem_size, n_actions), dtype=np.float32)
+        self.reward_memory = np.zeros(self.mem_size, dtype=np.float32)
         self.terminal_memory = np.zeros(self.mem_size, dtype=bool)
 
         self.device = device
